requirements.txt
flask==3.0.0
python-doctr[torch]==0.7.0
transformers==4.35.0
torch==2.1.0
Pillow==10.1.0
pdf2image==1.16.3
pandas==2.1.0
plotly==5.18.0